requirements.txt
pandas==2.2.2
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlalchemy==2.0.30
dbt-core
dbt-mysql
